+++ internal/cloud-storage/hash.go
