+++ internal/cloud-storage/remove.go
