+++ internal/codebase/hash.go
